refactor: controller test -> integration test로 변경
